requirements.txt
Flask==2.0.1
Flask-MySQLdb==0.2.0
requests==2.26.0
Werkzeug==2.2.2
psycopg2-binary
